AIPT/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
sqlmodel>=0.0.22
alembic==1.13.1
python-dotenv==1.0.1
pydantic>=2.5,<3
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0
httpx==0.26.0
ruff==0.2.1
black==24.1.1
typing_extensions>=4.9
